Drain script is redundant and stops proxy too soon
Secureproxy is deployed as a job in a BOSH release alongside the gorouter. When the gorouter drain script runs, it sets its health check to unhealthy, waits for longer than the downstream ALB health check period, then shuts down. We have observed secureproxy shutting down before the gorouter in production, resulting in 502s when the ALBs keep trying to reach it, since they haven't yet determined that it is unhealthy. This is because the secureproxy drain script immediately shuts the proxy down instead of keeping it up while the gorouter drains.

I considered modifying the script to wait for a configurable amount of time before draining, but then realized that the drain script isn't fulfilling its intended purpose: it's shutting the program down instead of draining it. Monit is already equipped to shut secureproxy down by calling `secureproxy_ctl stop`, which in turn does the same thing the drain script did. I removed the drain script and verified that the job can still shut down correctly -- now with no 502s returned.